🐎 Removes forced hardware acceleration on minimap
With the new view-aware rendering of the minimap a lot of time is now
passed refreshing the gpu textures when new lines are added/removed
during scroll. Using the profiling tools confirm that.

This commit includes removal of all the `accelerated` class inclusion
in container node and put it instead on lines.
It also includes the change of the translate method to generate
`translate` transformation instead of `translate3d`.