:bug: Fix smooth scrolling issue with visible area
As discovered by @jrouviere the callback `minimap:updated` was the
cause of the stop
before the edges.
After removing it the minimap lack a proper placement because it is
initialized before the tabs view affect the layout.
To solve that a mutation observer is created on the pane
and the minimap is relocated whenever children are added to or removed
from the pane view.
Finally the two layers surrounding the editor scroll at the same time
than the minimap.