:fire: Remove redundant tests
The size and position of the controls div is already tested above.
Testing position of the elements inside doesn’t add any value
and using absolute values in the tests when we know that both Chrome
flex box and TravisCI environment aren’t reliable enough to use this
kind of tests.