:fire: Remove screen delta guards
Screen delta represents the differences in number of lines between the
version before the change and the version after the change.
Changes that comes from the TextEditor always have a screen delta. The
only ones that doesn’t comes from the decoration management. And in
fact, manipulating decorations doesn’t changes the screen lines so
their screen delta is always 0.