:bug: Fix duplicated lines at the end of the minimap
This problem occurs when the only changes are screen lines changes due
to soft wrapping (by splitting panes, resizing the window, or just at
intialization). In that case, we don’t compute an extra intact range
that would copy a part of the offscreen buffer at the wrong place.

Fixes #292