:racehorse: Speed up scrolling with lots of decorations
Use case: Opening bootstrap.min.css with `minimap-pigments` enabled and
scrolling make the UI all laggy due to the permanent computation of the
decoration marker’s screen range.

Now we’ll cache the mapping type->rows->decorations for the whole
buffer and rebuild it each time a decoration is added/removed or a
marker change.