Separate decorations and tokens rendering between three layers
The main point of this optimisation is to avoid rendering the line
tokens whenever a decoration is added, removed or changed. The various
profiling I made in the past weeks all led to the `line.tokens` getter
being too costly.

So now, decorations are rendered on two layers (back and front) and
line tokens on the last one.